+++ guard/on/checked_oint_guard.ml
@@ -1,0 +1,5 @@
+type 'a t = Obj.t * 'a
+
+let wrap x = Obj.new_block Obj.abstract_tag 1, x
+
+let unwrap (_guard, x) = x